♻️Refactor: deal, person, detail 페이지 pos 디렉토리로 이동
