+++ src/pages/board/post/deal.css
@@ -6,7 +6,7 @@
 
 .deal-item {
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
   gap: 1.5rem;
   padding: 1.5rem;
   border-bottom: 1px solid #eee;
