+++ src/pages/board/post/deal.jsx
