+++ src/pages/board/post/detail.css
