+++ src/pages/board/post/person.jsx
@@ -21,7 +21,6 @@ export default function Person() {
       post_id: 2,
       title: '주말 레이드 동행 구합니다',
       price: 0,
-      type: '동행',
       image: 'https://via.placeholder.com/120',
       author: '작성자2',
       author_uuid: 'asdf',
@@ -33,7 +32,6 @@ export default function Person() {
       post_id: 3,
       title: 'PVP 대회 용병 구합니다',
       price: 0,
-      type: '용병',
       image: 'https://via.placeholder.com/120',
       author: '작성자3',
       author_uuid: 'zxcv',
